services/hf_elim_pubsub/requirements.txt
# HF Elimination Arbitrage — GCP Pub/Sub Service
flask>=3.0
gunicorn>=22.0
httpx>=0.27
loguru>=0.7
google-cloud-pubsub>=2.21
google-cloud-firestore>=2.16
python-dotenv>=1.0
